Study information card is now an importable component
